gateway/requirements.txt
fastapi
uvicorn[standard]
websockets
livekit-server-sdk
opuslib

